add compressed log and to use valkey_stat for container_logs
