+++ src/ai/backend/common/clients/valkey_client/valkey_stat/client.py
@@ -1061,3 +1061,80 @@ class ValkeyStatClient:
 
         if updates:
             await self.set_multiple_keys(updates)
+
+    @valkey_decorator()
+    async def enqueue_container_logs(
+        self,
+        container_id: str,
+        logs: bytes,
+    ) -> None:
+        """
+        Enqueue logs for a specific container.
+        TODO: Replace with a more efficient log storage solution.
+
+        :param container_id: The ID of the container.
+        :param logs: The logs to enqueue.
+        :raises: GlideClientError if the logs cannot be enqueued.
+        """
+        key = self._container_log_key(container_id)
+        tx = self._create_batch()
+        tx.rpush(
+            key,
+            [logs],
+        )
+        tx.expire(
+            key,
+            3600,  # 1 hour expiration
+        )
+        await self._client.client.exec(tx, raise_on_error=True)
+
+    @valkey_decorator()
+    async def container_log_len(
+        self,
+        container_id: str,
+    ) -> int:
+        """
+        Get the length of logs for a specific container.
+
+        :param container_id: The ID of the container.
+        :return: The number of logs for the container.
+        :raises: GlideClientError if the length cannot be retrieved.
+        """
+        key = self._container_log_key(container_id)
+        return await self._client.client.llen(key)
+
+    @valkey_decorator()
+    async def pop_container_logs(
+        self,
+        container_id: str,
+        count: int = 1,
+    ) -> Optional[list[bytes]]:
+        """
+        Pop logs for a specific container.
+
+        :param container_id: The ID of the container.
+        :return: List of logs for the container.
+        :raises: GlideClientError if the logs cannot be popped.
+        """
+        key = self._container_log_key(container_id)
+        return await self._client.client.lpop_count(key, count)
+
+    @valkey_decorator()
+    async def clear_container_logs(
+        self,
+        container_id: str,
+    ) -> None:
+        """
+        Clear logs for a specific container.
+
+        :param container_id: The ID of the container.
+        :raises: GlideClientError if the logs cannot be cleared.
+        """
+        key = self._container_log_key(container_id)
+        await self._client.client.delete([key])
+
+    def _container_log_key(self, container_id: str) -> str:
+        return f"containerlog.{container_id}"
+
+    def _create_batch(self, is_atomic: bool = False) -> Batch:
+        return Batch(is_atomic=is_atomic)
