+++ src/ai/backend/manager/event_dispatcher/dispatch.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 
+from ai.backend.common.clients.valkey_client.valkey_stat.client import ValkeyStatClient
 from ai.backend.common.clients.valkey_client.valkey_stream.client import ValkeyStreamClient
 from ai.backend.common.events.dispatcher import (
     CoalescingOptions,
@@ -91,6 +92,7 @@ from .reporters import EventLogger
 
 @dataclass
 class DispatcherArgs:
+    valkey_stat: ValkeyStatClient
     valkey_stream: ValkeyStreamClient
     scheduler_dispatcher: SchedulerDispatcher
     event_hub: EventHub
@@ -123,7 +125,10 @@ class Dispatchers:
         )
         self._image_event_handler = ImageEventHandler(args.agent_registry, args.db)
         self._kernel_event_handler = KernelEventHandler(
-            args.valkey_stream, args.agent_registry, args.db
+            args.valkey_stream,
+            args.valkey_stat,
+            args.agent_registry,
+            args.db,
         )
         self._schedule_event_handler = ScheduleEventHandler(args.scheduler_dispatcher)
         self._model_serving_event_handler = ModelServingEventHandler(args.agent_registry, args.db)
